pkp/pkp-lib#3002 exclude custom block manager plugin from the upgrade fix enabledSitePlugins

+++ classes/install/Upgrade.inc.php
@@ -2442,7 +2442,9 @@ class Upgrade extends Installer {
 		foreach ($allPlugins as $plugin) {
 			if ($plugin->isSitePlugin()) {
 				$pluginName = strtolower_codesafe($plugin->getName());
-				$result = $pluginSettings->update('DELETE FROM plugin_settings WHERE plugin_name = ? AND setting_name = \'enabled\' AND context_id <> 0', array($pluginName));
+				if ($pluginName != 'customblockmanagerplugin') {
+					$result = $pluginSettings->update('DELETE FROM plugin_settings WHERE plugin_name = ? AND setting_name = \'enabled\' AND context_id <> 0', array($pluginName));
+				}
 			}
 		}
